Change the Mem interface.
The core<->mem interface will use a wr[4:0] signal, to indicate which
byte will be used in the write operation (simplifies the memory module).
The cpath<->dpath interface will use a similar interface, but using
the typ[3:0] signal. The control unit should decode this signal, and
create the wr[4:0] signal (TODO).