Move CSR R/W operations to MEM stage.
All access to the CSR file register must be done in the MEM stage, to
avoid execution of operations in an out-of-orden sequence.
Complete detection of ECALL is done in MEM state too, due to the prv
dependency in the check.

Also, update the stall rules: Stall ID if the instruction at EX stage is
a valid CSR operation.