Upgrade the ALU with the mult and div units.
This is the easy way to implement the RV32M ISA: adding the multiplier
and divider unit to the ALU.
Extend the ALU opcodes, and the data mux.
Remove previous attempt in the dpath (wrong commit).
Update the control signals (shift/extend with 1-bit due to the ALU
opcode extention).
Update the instruction decoding.
Update stall logic: full stall with mult/div instructions.